+++ src/types.ts
@@ -1,14 +1,17 @@
-import { ReactElement } from 'react';
+import { ReactElement, MutableRefObject } from 'react';
+import Swiper, {
+  SwiperOptions,
+  SelectableElement as SwiperSelectableElement,
+  SwiperModule
+} from 'swiper';
 
-import Swiper, { SwiperOptions, SelectableElement, SwiperModule } from 'swiper';
+type Maybe<T> = T | null;
 
-export type ReactIdSwiperRenderProps = (props: ReactIdSwiperProps) => ReactElement | null;
+export type ReactIdSwiperRenderProps = (props: ReactIdSwiperProps) => Maybe<ReactElement>;
 
 export type WrappedElementType = 'div' | 'section' | 'span';
 
 export type ReactIdSwiperChildren = ReactElement | ReactElement[];
-
-export type GetSwiper = (swiper: SwiperInstance) => void;
 
 export type SwiperModules = (SwiperModule & { name: string })[];
 
@@ -20,14 +23,13 @@ export interface ReactIdSwiperProps extends SwiperOptions {
   slideClass?: string;
   rebuildOnUpdate?: boolean;
   shouldSwiperUpdate?: boolean;
-  getSwiper?: GetSwiper;
   activeSlideKey?: string;
   renderScrollbar?: ReactIdSwiperRenderProps;
   renderPagination?: ReactIdSwiperRenderProps;
   renderPrevButton?: ReactIdSwiperRenderProps;
   renderNextButton?: ReactIdSwiperRenderProps;
   renderParallax?: ReactIdSwiperRenderProps;
-  rtl?: string | undefined;
+  rtl?: string;
   children?: ReactIdSwiperChildren;
   parallaxEl?: {
     el: string;
@@ -40,9 +42,19 @@ export interface ReactIdSwiperCustomProps extends ReactIdSwiperProps {
   Swiper: typeof Swiper;
 }
 
-export type SelectableElement = SelectableElement | undefined;
+export interface SwiperRefNode extends HTMLDivElement {
+  swiper?: SwiperInstance;
+}
 
-export type SwiperInstance = Swiper | null;
+export type SelectableElement = SwiperSelectableElement | undefined;
+
+export type SwiperInstance = Maybe<Swiper>;
+
+export type RefType<T> = Maybe<((instance: Maybe<T>) => void) | MutableRefObject<Maybe<T>>>;
+
+export type SetRef = <V>(ref: RefType<V>, value: V) => void;
+
+export type UseForkRef = <T>(refA: RefType<T>, refB: RefType<T>) => Maybe<(v: T) => void>;
 
 export type SwiperModuleName =
   | 'navigation'