+++ src/utils.ts
@@ -1,13 +1,15 @@
-import { Children, isValidElement, ReactElement } from 'react';
-
-import { ReactIdSwiperChildren, SelectableElement, SwiperModuleName, SwiperModules } from './types';
+import { isValidElement, Children, ReactElement } from 'react';
+import {
+  SelectableElement,
+  SwiperModules,
+  SwiperModuleName,
+  ReactIdSwiperChildren,
+  SetRef
+} from './types';
 
 export const classNames = (el: SelectableElement): string => {
   if (typeof el === 'string') {
-    return el
-      .split('.')
-      .join(' ')
-      .trim();
+    return el.split('.').join(' ').trim();
   } else if (el instanceof HTMLElement) {
     return el.className;
   }
@@ -15,7 +17,7 @@ export const classNames = (el: SelectableElement): string => {
   return '';
 };
 
-export const validateChildren = (children: ReactIdSwiperChildren) => {
+export const validateChildren = (children: ReactIdSwiperChildren): boolean => {
   let isValid = true;
 
   if (Array.isArray(children)) {
@@ -46,4 +48,12 @@ export const isModuleAvailable = (modules: SwiperModules, moduleName: SwiperModu
   }
 
   return moduleAvailable;
+};
+
+export const setRef: SetRef = (ref, value) => {
+  if (typeof ref === 'function') {
+    ref(value);
+  } else if (ref) {
+    ref.current = value;
+  }
 };